requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
Pillow>=10.3.0
numpy>=1.26.0
torch>=2.1.0
transformers>=4.39.0
huggingface_hub>=0.20.0
safetensors>=0.4.2
timm
hf_xet
pytest
